requirements.txt
aria2p==0.12.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
beautifulsoup4>=4.12.2
selenium>=4.0.0